Move XStream processor from salesforce to a dedicated extension #34
