+++ bom/pom.xml
@@ -174,6 +174,11 @@
                 <artifactId>camel-quarkus-jetty-common</artifactId>
                 <version>${project.version}</version>
             </dependency>
+            <dependency>
+                <groupId>org.apache.camel.quarkus</groupId>
+                <artifactId>camel-quarkus-xstream-common</artifactId>
+                <version>${project.version}</version>
+            </dependency>
 
             <!-- Quarkus Camel -->
             <dependency>
