+++ build-parent/pom.xml
@@ -37,6 +37,7 @@
 
         <groovy.version>2.5.7</groovy.version>
         <jetty.version>9.4.18.v20190429</jetty.version>
+        <xstream.version>1.4.11</xstream.version>
 
         <maven-checkstyle-plugin.version>3.0.0</maven-checkstyle-plugin.version>
         <maven-checkstyle.version>7.6.1</maven-checkstyle.version>
@@ -75,6 +76,12 @@
             <dependency>
                 <groupId>org.apache.camel.quarkus</groupId>
                 <artifactId>camel-quarkus-jetty-common-deployment</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+
+            <dependency>
+                <groupId>org.apache.camel.quarkus</groupId>
+                <artifactId>camel-quarkus-xstream-common-deployment</artifactId>
                 <version>${project.version}</version>
             </dependency>
