+++ extensions/xstream-common/pom.xml
@@ -17,37 +17,25 @@
     limitations under the License.
 
 -->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 
-    <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>org.apache.camel.quarkus</groupId>
-        <artifactId>camel-quarkus-parent</artifactId>
+        <artifactId>camel-quarkus-build-parent</artifactId>
         <version>0.0.1-SNAPSHOT</version>
+        <relativePath>../../build-parent/pom.xml</relativePath>
     </parent>
 
-    <artifactId>camel-quarkus-extensions</artifactId>
+    <modelVersion>4.0.0</modelVersion>
+
+    <artifactId>camel-quarkus-xstream-common-parent</artifactId>
+    <name>Camel Quarkus :: XStream :: Common</name>
     <packaging>pom</packaging>
-
-    <name>Camel Quarkus :: Extensions</name>
-
     <modules>
-        <module>jetty-common</module>
-        <module>xstream-common</module>
-        <module>netty4-http</module>
-        <module>core</module>
-        <module>infinispan</module>
-        <module>aws-s3</module>
-        <module>aws-sns</module>
-        <module>aws-sqs</module>
-        <module>jdbc</module>
-        <module>salesforce</module>
-        <module>servlet</module>
-        <module>log</module>
-        <module>direct</module>
-        <module>rest</module>
-        <module>timer</module>
+        <module>deployment</module>
+        <module>runtime</module>
     </modules>
 
 </project>
